agent/requirements.txt
requests>=2.31.0
psutil>=5.9.0
WMI>=1.5.1
pywin32>=306
PyYAML>=6.0

